+++ das_client/app/integration_test/test/automatic_advancement_test.dart
@@ -1,6 +1,8 @@
+import 'package:app/di/di.dart';
 import 'package:app/pages/journey/train_journey/widgets/header/header.dart';
 import 'package:app/pages/journey/train_journey/widgets/header/start_pause_button.dart';
 import 'package:app/pages/journey/train_journey/widgets/table/cells/route_chevron.dart';
+import 'package:app/util/time_constants.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -52,8 +54,10 @@ void main() {
       await tester.dragUntilVisible(find.text('B1'), scrollableFinder, const Offset(0, 100));
       expect(find.text('Bern'), findsAny);
 
-      // Wait until idle time reached
-      await Future.delayed(const Duration(seconds: 12));
+      final waitTime = DI.get<TimeConstants>().automaticAdvancementIdleTimeAutoScroll + 1;
+
+      // wait until waitTime reached
+      await Future.delayed(Duration(seconds: waitTime));
       await tester.pumpAndSettle();
 
       // Check if the last row is visible